Support accessing image tile by location in x/y
The use of `tiledImage = {}` rather than an array is because
Python complains a bunch when you try to access an array like
that. The one issue with this implementation is that the tile
cannot be directly shown from the matrix dictionary, instead
requiring a temporary variable to allow object methods to
function.